refactor(elysia): match red's nested layout (no session/guard files)
Align Phase 2 orange modules with the structure red landed on PR #882:

  auth/{index,model,status,login,logout}.ts
  settings/{index,model,get,update}.ts
  feed/{index,model,list,create}.ts

- Drop auth/session.ts — helpers live in auth/index.ts alongside the
  composed authRoutes so settings/feed can import them directly.
- Drop settings/guard.ts and feed/guard.ts — each group's onBeforeHandle
  lives inline in its index.ts.
- Rename feed/append.ts → feed/create.ts.
- model.ts files hold only TypeBox schemas; added a comment on LoginBody
  explaining the t.Optional(t.String()) choice (contract test expects
  { success: false, error: "Password required" } rather than Elysia's
  default 422 validation shape).
- No explicit Static<> type annotations — handlers rely on Elysia's
  inference from the route schema.

bun run build (tsc --noEmit) passes. All files 9–102 lines.
