+++ src/routes-elysia/feed/list.ts
@@ -1,79 +1,76 @@
 import { Elysia } from 'elysia';
 import fs from 'fs';
 import { FEED_LOG } from '../../config.ts';
-import { authGuard } from './guard.ts';
 import { FeedQuery, type FeedEvent } from './model.ts';
 
 const MAW_JS_URL = process.env.MAW_JS_URL || 'http://localhost:3456';
 
-export const listFeedRoute = new Elysia()
-  .use(authGuard)
-  .get('/', async ({ query, set }) => {
-    try {
-      const limit = Math.min(200, parseInt(query.limit || '50'));
-      const oracle = query.oracle || undefined;
-      const event = query.event || undefined;
-      const since = query.since || undefined;
+export const listFeedRoute = new Elysia().get('/', async ({ query, set }) => {
+  try {
+    const limit = Math.min(200, parseInt(query.limit || '50'));
+    const oracle = query.oracle || undefined;
+    const event = query.event || undefined;
+    const since = query.since || undefined;
 
-      let allEvents: FeedEvent[] = [];
+    let allEvents: FeedEvent[] = [];
 
-      if (fs.existsSync(FEED_LOG)) {
-        const raw = fs.readFileSync(FEED_LOG, 'utf-8').trim().split('\n').filter(Boolean);
-        const localEvents: FeedEvent[] = raw.map(line => {
-          const [ts, oracleName, host, eventType, project, rest] = line.split(' | ').map(s => s.trim());
-          const [sessionId, ...msgParts] = (rest || '').split(' » ');
-          return {
-            timestamp: ts,
-            oracle: oracleName,
-            host,
-            event: eventType,
-            project,
-            session_id: sessionId?.trim(),
-            message: msgParts.join(' » ').trim(),
-            source: 'local',
-          };
-        });
-        allEvents.push(...localEvents);
-      }
-
-      try {
-        const mawRes = await fetch(`${MAW_JS_URL}/api/feed?limit=100`, { signal: AbortSignal.timeout(2000) });
-        if (mawRes.ok) {
-          const mawData = await mawRes.json() as any;
-          if (mawData.events && Array.isArray(mawData.events)) {
-            const mawEvents: FeedEvent[] = mawData.events.map((e: any) => ({
-              timestamp: e.timestamp || new Date(e.ts).toISOString().replace('T', ' ').slice(0, 19),
-              oracle: e.oracle,
-              host: e.host,
-              event: e.event,
-              project: e.project,
-              session_id: e.sessionId,
-              message: e.message,
-              source: 'maw-js',
-            }));
-            allEvents.push(...mawEvents);
-          }
-        }
-      } catch (mawError) {
-        console.log('maw-js feed unavailable:', mawError);
-      }
-
-      if (oracle) allEvents = allEvents.filter(e => e.oracle === oracle);
-      if (event) allEvents = allEvents.filter(e => e.event === event);
-      if (since) allEvents = allEvents.filter(e => e.timestamp >= since);
-
-      allEvents.sort((a, b) => new Date(b.timestamp).getTime() - new Date(a.timestamp).getTime());
-      const total = allEvents.length;
-      allEvents = allEvents.slice(0, limit);
-
-      const fiveMinAgo = new Date(Date.now() - 5 * 60_000).toISOString().replace('T', ' ').slice(0, 19);
-      const activeOracles = [...new Set(allEvents.filter(e => e.timestamp >= fiveMinAgo).map(e => e.oracle))];
-
-      return { events: allEvents, total, active_oracles: activeOracles };
-    } catch (e: any) {
-      set.status = 500;
-      return { error: e.message, events: [], total: 0 };
+    if (fs.existsSync(FEED_LOG)) {
+      const raw = fs.readFileSync(FEED_LOG, 'utf-8').trim().split('\n').filter(Boolean);
+      const localEvents: FeedEvent[] = raw.map(line => {
+        const [ts, oracleName, host, eventType, project, rest] = line.split(' | ').map(s => s.trim());
+        const [sessionId, ...msgParts] = (rest || '').split(' » ');
+        return {
+          timestamp: ts,
+          oracle: oracleName,
+          host,
+          event: eventType,
+          project,
+          session_id: sessionId?.trim(),
+          message: msgParts.join(' » ').trim(),
+          source: 'local',
+        };
+      });
+      allEvents.push(...localEvents);
     }
-  }, {
-    query: FeedQuery,
-  });
+
+    try {
+      const mawRes = await fetch(`${MAW_JS_URL}/api/feed?limit=100`, { signal: AbortSignal.timeout(2000) });
+      if (mawRes.ok) {
+        const mawData = await mawRes.json() as any;
+        if (mawData.events && Array.isArray(mawData.events)) {
+          const mawEvents: FeedEvent[] = mawData.events.map((e: any) => ({
+            timestamp: e.timestamp || new Date(e.ts).toISOString().replace('T', ' ').slice(0, 19),
+            oracle: e.oracle,
+            host: e.host,
+            event: e.event,
+            project: e.project,
+            session_id: e.sessionId,
+            message: e.message,
+            source: 'maw-js',
+          }));
+          allEvents.push(...mawEvents);
+        }
+      }
+    } catch (mawError) {
+      console.log('maw-js feed unavailable:', mawError);
+    }
+
+    if (oracle) allEvents = allEvents.filter(e => e.oracle === oracle);
+    if (event) allEvents = allEvents.filter(e => e.event === event);
+    if (since) allEvents = allEvents.filter(e => e.timestamp >= since);
+
+    allEvents.sort((a, b) => new Date(b.timestamp).getTime() - new Date(a.timestamp).getTime());
+    const total = allEvents.length;
+    allEvents = allEvents.slice(0, limit);
+
+    const fiveMinAgo = new Date(Date.now() - 5 * 60_000).toISOString().replace('T', ' ').slice(0, 19);
+    const activeOracles = [...new Set(allEvents.filter(e => e.timestamp >= fiveMinAgo).map(e => e.oracle))];
+
+    return { events: allEvents, total, active_oracles: activeOracles };
+  } catch (e: any) {
+    set.status = 500;
+    return { error: e.message, events: [], total: 0 };
+  }
+}, {
+  query: FeedQuery,
+});
