--- src/routes-elysia/settings/guard.ts
@@ -1,11 +1,0 @@
-import { Elysia } from 'elysia';
-import { SESSION_COOKIE_NAME, isAuthenticated } from '../auth/session.ts';
-
-export const authGuard = new Elysia({ name: 'settings-auth-guard' })
-  .onBeforeHandle(({ server, request, cookie, set }) => {
-    const sessionValue = cookie[SESSION_COOKIE_NAME]?.value as string | undefined;
-    if (!isAuthenticated(server, request, sessionValue)) {
-      set.status = 401;
-      return { error: 'Unauthorized', requiresAuth: true };
-    }
-  });
